convertPlanilha/requirements.txt
botcity-framework-web>=0.8.0,<1.0
botcity-maestro-sdk>=0.3.3,<1.0
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
# pandas

requests